+++ src/nativeOptions.ts
@@ -1,9 +1,14 @@
 import { Capacitor } from '@capacitor/core';
 import type { BrowserOptions } from '@sentry/browser';
 import type { CapacitorOptions } from './options';
+import { getCurrentSpotlightUrl } from './utils/webViewUrl';
 
 interface CapacitorLoggerOptions {
   enableLogs: boolean
+}
+
+interface CapacitorSpotlightOptions {
+  sidecarUrl: string;
 }
 
 /**
@@ -41,7 +46,8 @@ export function FilterNativeOptions(
     // tunnel: options.tunnel: Only handled on the JavaScript Layer.
     enableCaptureFailedRequests: options.enableCaptureFailedRequests,
     ...iOSParameters(options),
-    ...LogParameters(options)
+    ...LogParameters(options),
+    ...SpotlightParameters(),
   };
 }
 
@@ -60,6 +66,14 @@ function LogParameters(options: CapacitorOptions & BrowserOptions ): CapacitorLo
   // Only Web and Android implements log parameters initialization.
   if (shouldEnable && Capacitor.getPlatform() === 'android') {
     return { enableLogs: shouldEnable };
+  }
+  return undefined;
+}
+
+function SpotlightParameters(): CapacitorSpotlightOptions | undefined {
+  const spotlightUrl = getCurrentSpotlightUrl();
+  if (spotlightUrl) {
+    return { sidecarUrl: spotlightUrl };
   }
   return undefined;
 }